refactor: migrate 3 types to adminportal structure
