+++ docs/docs/auto-docs/types/AdminPortal/membership/type-aliases/MembershipRequest.md
@@ -1,0 +1,33 @@
+[Admin Docs](/)
+
+***
+
+# Type Alias: MembershipRequest
+
+> **MembershipRequest** = `object`
+
+Defined in: [src/types/AdminPortal/membership.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/membership.ts#L4)
+
+## Properties
+
+### \_id
+
+> **\_id**: `string`
+
+Defined in: [src/types/AdminPortal/membership.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/membership.ts#L5)
+
+***
+
+### organization
+
+> **organization**: [`Organization`](../../Organization/type/type-aliases/Organization.md)
+
+Defined in: [src/types/AdminPortal/membership.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/membership.ts#L6)
+
+***
+
+### user
+
+> **user**: [`User`](../../../shared-components/User/type/type-aliases/User.md)
+
+Defined in: [src/types/AdminPortal/membership.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/membership.ts#L7)
