+++ docs/docs/auto-docs/types/AdminPortal/pagination/type-aliases/DefaultConnectionPageInfo.md
@@ -1,0 +1,41 @@
+[Admin Docs](/)
+
+***
+
+# Type Alias: DefaultConnectionPageInfo
+
+> **DefaultConnectionPageInfo** = `object`
+
+Defined in: [src/types/AdminPortal/pagination.ts:1](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/pagination.ts#L1)
+
+## Properties
+
+### endCursor?
+
+> `optional` **endCursor**: `string`
+
+Defined in: [src/types/AdminPortal/pagination.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/pagination.ts#L5)
+
+***
+
+### hasNextPage
+
+> **hasNextPage**: `boolean`
+
+Defined in: [src/types/AdminPortal/pagination.ts:2](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/pagination.ts#L2)
+
+***
+
+### hasPreviousPage
+
+> **hasPreviousPage**: `boolean`
+
+Defined in: [src/types/AdminPortal/pagination.ts:3](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/pagination.ts#L3)
+
+***
+
+### startCursor?
+
+> `optional` **startCursor**: `string`
+
+Defined in: [src/types/AdminPortal/pagination.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/pagination.ts#L4)
