--- docs/docs/auto-docs/types/address/type-aliases/Address.md
@@ -1,73 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Type Alias: Address
-
-> **Address** = `object`
-
-Defined in: [src/types/address.ts:1](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L1)
-
-## Properties
-
-### city?
-
-> `optional` **city**: `string`
-
-Defined in: [src/types/address.ts:2](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L2)
-
-***
-
-### countryCode?
-
-> `optional` **countryCode**: `string`
-
-Defined in: [src/types/address.ts:3](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L3)
-
-***
-
-### dependentLocality?
-
-> `optional` **dependentLocality**: `string`
-
-Defined in: [src/types/address.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L4)
-
-***
-
-### line1?
-
-> `optional` **line1**: `string`
-
-Defined in: [src/types/address.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L5)
-
-***
-
-### line2?
-
-> `optional` **line2**: `string`
-
-Defined in: [src/types/address.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L6)
-
-***
-
-### postalCode?
-
-> `optional` **postalCode**: `string`
-
-Defined in: [src/types/address.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L7)
-
-***
-
-### sortingCode?
-
-> `optional` **sortingCode**: `string`
-
-Defined in: [src/types/address.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L8)
-
-***
-
-### state?
-
-> `optional` **state**: `string`
-
-Defined in: [src/types/address.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L9)
