--- docs/docs/auto-docs/types/address/type-aliases/AddressInput.md
@@ -1,73 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Type Alias: AddressInput
-
-> **AddressInput** = `object`
-
-Defined in: [src/types/address.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L12)
-
-## Properties
-
-### city?
-
-> `optional` **city**: `string`
-
-Defined in: [src/types/address.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L13)
-
-***
-
-### countryCode?
-
-> `optional` **countryCode**: `string`
-
-Defined in: [src/types/address.ts:14](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L14)
-
-***
-
-### dependentLocality?
-
-> `optional` **dependentLocality**: `string`
-
-Defined in: [src/types/address.ts:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L15)
-
-***
-
-### line1?
-
-> `optional` **line1**: `string`
-
-Defined in: [src/types/address.ts:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L16)
-
-***
-
-### line2?
-
-> `optional` **line2**: `string`
-
-Defined in: [src/types/address.ts:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L17)
-
-***
-
-### postalCode?
-
-> `optional` **postalCode**: `string`
-
-Defined in: [src/types/address.ts:18](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L18)
-
-***
-
-### sortingCode?
-
-> `optional` **sortingCode**: `string`
-
-Defined in: [src/types/address.ts:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L19)
-
-***
-
-### state?
-
-> `optional` **state**: `string`
-
-Defined in: [src/types/address.ts:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/address.ts#L20)
