--- docs/docs/auto-docs/types/membership/type-aliases/MembershipRequest.md
@@ -1,33 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Type Alias: MembershipRequest
-
-> **MembershipRequest** = `object`
-
-Defined in: [src/types/membership.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/membership.ts#L4)
-
-## Properties
-
-### \_id
-
-> **\_id**: `string`
-
-Defined in: [src/types/membership.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/membership.ts#L5)
-
-***
-
-### organization
-
-> **organization**: [`Organization`](../../AdminPortal/Organization/type/type-aliases/Organization.md)
-
-Defined in: [src/types/membership.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/membership.ts#L6)
-
-***
-
-### user
-
-> **user**: [`User`](../../shared-components/User/type/type-aliases/User.md)
-
-Defined in: [src/types/membership.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/membership.ts#L7)
