+++ src/types/AdminPortal/Organization/type.ts
@@ -1,8 +1,8 @@
-import type { Address, AddressInput } from 'types/address';
+import type { Address, AddressInput } from 'types/AdminPortal/address';
 import type { User } from 'types/shared-components/User/type';
 import type { ActionItemCategory } from 'types/AdminPortal/actionItem';
 import type { AgendaCategory } from 'types/Agenda/type';
-import type { MembershipRequest } from 'types/membership';
+import type { MembershipRequest } from 'types/AdminPortal/membership';
 import type { Post } from 'types/Post/type';
 import type { Venue } from 'types/AdminPortal/venue';
