+++ src/types/AdminPortal/membership.ts
@@ -1,4 +1,4 @@
-import type { User } from './shared-components/User/type';
+import type { User } from '../shared-components/User/type';
 import type { Organization } from 'types/AdminPortal/Organization/type';
 
 export type MembershipRequest = {
